+++ examples/python/empirical_transition_matrix.py
@@ -25,12 +25,13 @@ import pandas as pd
 import transitionMatrix as tm
 from transitionMatrix import source_path
 from transitionMatrix.estimators import aalen_johansen_estimator as aj
+from transitionMatrix.utils.converters import datetime_to_float
 
 dataset_path = source_path + "datasets/"
 
-
 # Example 1: Credit Rating Migration example
 # Example 2: Simple 2x2 Matrix for testing
+# Example 3: Credit Rating Migration example with timestamps in raw date format
 
 example = 1
 
@@ -44,16 +45,23 @@ if example == 1:
     data = pd.read_csv(dataset_path + 'synthetic_data7.csv', dtype={'State': str})
 elif example == 2:
     data = pd.read_csv(dataset_path + 'synthetic_data8.csv', dtype={'State': str})
+elif example == 3:
+    data = pd.read_csv(dataset_path + 'synthetic_data9.csv', parse_dates=True)
+    # convert datetime data to floats, return also the observation window data
+    bounds, data = datetime_to_float(data)
+    print('Start and End dates', bounds)
+
 sorted_data = data.sort_values(['Time', 'ID'], ascending=[True, True])
+print(sorted_data.head(5))
 print(sorted_data.describe())
 
 # Step 2
 # Describe and validate the State Space against the data
 print("> Step 2: Describe and validate the State Space against the data")
 # We insert the expected labels of the state space
-if example == 1:
+if example == 1 or example == 3:
     definition = [('0', "AAA"), ('1', "AA"), ('2', "A"), ('3', "BBB"),
-                   ('4', "BB"), ('5', "B"), ('6', "CCC"), ('7', "D")]
+                  ('4', "BB"), ('5', "B"), ('6', "CCC"), ('7', "D")]
 elif example == 2:
     definition = [('0', "G"), ('1', "B")]
 myState = tm.StateSpace(definition)
@@ -78,14 +86,12 @@ print(etm[:, :, -1])
 
 # Step 5
 # Create a visualization of the transition rates
-if example == 1:
+if example == 1 or example == 3:
     # Now lets plot a collection of curves for all ratings
     print("> Plot the transition curves")
 
     Periods = 10
     Ratings = 8
-
-    periods = range(0, Periods)
 
     m = 4
     n = 2
@@ -101,15 +107,15 @@ if example == 1:
         for rf in range(0, Ratings):
             cPD = etm[ri, rf, :]
             curves.append(cPD)
-            axarr[axj, axi].set_aspect(5)
+            # axarr[axj, axi].set_aspect(5)
             axarr[axj, axi].set_ylabel('State ' + str(ri), fontsize=12)
             axarr[axj, axi].set_xlabel("Time")
             axarr[axj, axi].plot(times[1:], curves[rf], label="RI=%d" % (rf,))
-            axarr[axj, axi].set_xticks(range(10), minor=False)
-            axarr[axj, axi].set_yticks(np.linspace(0,1,5), minor=False)
+            # axarr[axj, axi].set_xticks(range(10), minor=False)
+            axarr[axj, axi].set_yticks(np.linspace(0, 1, 5), minor=False)
             # axarr[axj, axi].yaxis.grid(True, which='minor')
             axarr[axj, axi].margins(y=0.05, x=0.05)
-            # axarr[axj, axi].autoscale()
+            axarr[axj, axi].autoscale()
             axarr[axj, axi].grid(True)
 
     # plt.tight_layout()